Added doc string for SetAssociativeCache structure and implemented SetAssociativeCache.size()

+++ src/set_associative_cache.h
@@ -3,6 +3,61 @@
 
 #include <cstdint>
 
+/**
+ * represents a set-associative cache
+ *
+ * 1-way set associative
+ * 8 sets, 1 line per set
+ *   +--------+
+ * 0 |        |
+ *   +--------+
+ * 1 |        |
+ *   +--------+
+ * 2 |        |
+ *   +--------+
+ * 3 |        |
+ *   +--------+
+ * 4 |        |
+ *   +--------+
+ * 5 |        |
+ *   +--------+
+ * 6 |        |
+ *   +--------+
+ * 7 |        |
+ *   +--------+
+ *
+ * 2-way set associative
+ * 4 sets, 2 lines per set
+ *   +--------+
+ * 0 |        |
+ *   |        |
+ *   +--------+
+ * 1 |        |
+ *   |        |
+ *   +--------+
+ * 2 |        |
+ *   |        |
+ *   +--------+
+ * 3 |        |
+ *   |        |
+ *   +--------+
+ *
+ * 4-way set associative
+ * 2 sets, 4 lines per set
+ *   +--------+
+ * 0 |        |
+ *   |        |
+ *   |        |
+ *   |        |
+ *   +--------+
+ * 1 |        |
+ *   |        |
+ *   |        |
+ *   |        |
+ *   +--------+
+ *
+ *   ways: is the number of lines per set
+ */
 class SetAssociativeCache {
 public:
     SetAssociativeCache(bool write_allocate, bool write_back, uint32_t miss_latency,
@@ -17,7 +72,7 @@ public:
     uint32_t sets_;
     uint32_t ways_;
 
-    uint32_t size();
+    uint64_t size();
 };
 
 #endif
